backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.3.0
numpy==1.26.4
